troubleshooting

+++ src/main/java/com/codeup/adlister/dao/MySQLReviewsDao.java
@@ -3,10 +3,7 @@ package com.codeup.adlister.dao;
 import com.codeup.adlister.models.Review;
 import com.mysql.cj.jdbc.Driver;
 
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.ResultSet;
-import java.sql.SQLException;
+import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -43,11 +40,11 @@ public class MySQLReviewsDao implements Reviews {
         try {
             String insertQuery = "INSERT INTO reviews(user_id, review_text, rating, image_text, ad_id) VALUES (?, ?, ?, ?, ?)";
             PreparedStatement stmt = connection.prepareStatement(insertQuery, Statement.RETURN_GENERATED_KEYS);
-            stmt.setInteger(1, review.getUserId());
+            stmt.setLong(1, review.getUserId());
             stmt.setString(2, review.getReviewText());
-            stmt.setInteger(3, review.getRating());
+            stmt.setInt(3, review.getRating());
             stmt.setString(4, review.getImageText()); //would this be 4 or 5 because of blob
-            stmt.setInteger(5, review.getAdId()); //5 or 6 because of blob
+            stmt.setLong(5, review.getAdId()); //5 or 6 because of blob
             stmt.executeUpdate();
             ResultSet rs = stmt.getGeneratedKeys();
             rs.next();
